// src/types/api.ts
export interface components {
    schemas: {
// ...
        ChatMessageResponse: {
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Role */
            role: string;
            /** Message */
            message: string;
            /**
             * Timestamp
             * Format: date-time
             */
            timestamp: string;
            /** Parent Id */
            parent_id?: string | null;
            /** Provider */
            provider?: string | null;
            /** Model */
            model?: string | null;
            /** Embedding Model */
            embedding_model?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}